Split Backdrop into multiple files
File started to become too long.
