+++ entry_types/scrolled/package/src/frontend/Backdrop/OrientationAwareBackgroundImage.js
@@ -1,0 +1,23 @@
+import React from 'react';
+
+import {usePortraitOrientation} from './../usePortraitOrientation';
+import {BackgroundImage} from './BackgroundImage';
+
+export function OrientationAwareBackgroundImage({image, onMotifAreaUpdate, imageMobile, containerDimension}) {
+  const mobile = usePortraitOrientation();
+
+  if (mobile) {
+    return (
+      <BackgroundImage image={imageMobile}
+                       onMotifAreaUpdate={onMotifAreaUpdate}
+                       containerDimension={containerDimension} />
+    )
+  }
+  else {
+    return (
+      <BackgroundImage image={image}
+                       onMotifAreaUpdate={onMotifAreaUpdate}
+                       containerDimension={containerDimension} />
+    )
+  }
+}
